move scripts to sagemaker folder in hardware
