+++ runhouse/resources/hardware/__init__.py
@@ -6,7 +6,7 @@ from .cluster_factory import (
     sagemaker_cluster,
 )
 from .on_demand_cluster import OnDemandCluster
-from .sagemaker_cluster import SageMakerCluster
+from .sagemaker.sagemaker_cluster import SageMakerCluster
 from .utils import (
     _current_cluster,
     _get_cluster_from,
