+++ runhouse/resources/hardware/sagemaker/refresh-ssm-session.sh
