+++ runhouse/resources/hardware/sagemaker/start-ssm-proxy-connection.sh
